# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CloudHub Storage Gateway API
  description: |
    CloudHub is a self-hosted cloud storage gateway that aggregates multiple 
    cloud storage providers into a single virtual storage pool.
    
    ## Features
    - Multi-provider support (Google Drive, Mega, OneDrive, Dropbox, R2, S3, B2, WebDAV)
    - Virtual filesystem with folder support
    - File upload, download, rename, delete, search
    - Transfer logging with retry mechanism
    - Rate limiting per user
    - Health check and capacity refresh workers
    
  version: 1.0.0
  contact:
    name: CloudHub Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
...
  /files:
    get:
      tags: [Files]
      summary: List files in a directory
      security:
        - bearerAuth: []
      parameters:
        - name: parent_id
          in: query
          schema:
            type: string
            format: uuid
          description: Parent folder ID (omit for root)
      responses:
        '200':
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileResponse'
...
  /files/{fileID}:
    get:
      tags: [Files]
      summary: Get file metadata
      security:
        - bearerAuth: []
      parameters:
        - name: fileID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: File metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Files]
      summary: Delete a file
      security:
        - bearerAuth: []
      parameters:
        - name: fileID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: File deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FileResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        virtual_path:
          type: string
        size:
          type: integer
        checksum:
          type: string
        mime_type:
          type: string
        is_directory:
          type: boolean
        parent_id:
          type: string
          format: uuid
          nullable: true
        location_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        code:
          type: integer
        status:
          type: string
        message:
          type: string
        details:
          type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...